+++ spec/services/test_hook_service_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe TestHookService do
+describe Hooks::Test do
   let (:user)    { create :user }
   let (:project) { create :project }
   let (:hook)    { create :project_hook, project: project }
@@ -8,7 +8,8 @@ describe TestHookService do
   describe :execute do
     it "should execute successfully" do
       stub_request(:post, hook.url).to_return(status: 200)
-      TestHookService.new.execute(hook, user).should be_true
+      result = Hooks::Test.perform(hook: hook, user: user)
+      result.should be_success
     end
   end
 end
